bug & style fix in examples++-load

+++ examples++-load/libff-mmap-semaphore.c
@@ -300,7 +300,7 @@ long ffmmap_read (ff_Pmmap p, void *pt, size_t ln, long off) {
 	void *pk = (char *)p->map + off;
 	memcpy(pt, pk, ln);
 	long *pp = (long *)pt;
-	if (ff_mmap_sem_verb > 9) {printf(" R %ld %ld %lu %p\n", *pp, off, ln, pk);}
+	if (ff_mmap_sem_verb > 9) {printf(" R %ld %ld %lu %p\n", *pp, off, (unsigned long)ln, pk);}
 
 	return ln;
 }
@@ -319,7 +319,7 @@ long ffmmap_write (ff_Pmmap p, void *pt, size_t ln, long off) {
 	void *pk = (char *)p->map + off;
 	memcpy(pk, pt, ln);
 	long *pp = (long *)pk;
-	if (ff_mmap_sem_verb > 9) {printf(" W %ld %ld %lu %p\n", *pp, off, ln, pk);}
+	if (ff_mmap_sem_verb > 9) {printf(" W %ld %ld %lu %p\n", *pp, off, (unsigned int)ln, pk);}
 
 	return ln;
 }
